+++ web/vite.config.ts
@@ -21,6 +21,10 @@ export default defineConfig({
     define: {
         __APP_VERSION__: JSON.stringify(localVersion),
         __APP_RELEASES__: JSON.stringify(parseChangelog(localChangelog)),
+        global: "globalThis",
+    },
+    optimizeDeps: {
+        include: ["ali-oss"],
     },
     server: {
         proxy: {